+++ CHANGELOG.md
@@ -41,6 +41,7 @@ Versioning follows [Semantic Versioning](https://semver.org/).
 - Department avatars on the home page now show actual member photos
 - "SU:Core" admin panel nav renamed to "Kanban" to avoid confusion
 - Removed redundant information from forms page
+- Forms viewer and forms builder pages on admin panel now merged into one manage questionnaires page, with a list of all questionnaires and a "Create a questionnaire" button
 
 ## [1.0.0] — 2026-06-20